requirements.txt
azure-functions==1.23.0
certifi==2025.4.26
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
